telegram_timesheet_bot/requirements.txt
numpy<2
fastapi==0.101.1
uvicorn==0.22.0
gspread==5.9.0
oauth2client==4.1.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
easyocr==1.6.2
# Note: EasyOCR depends on PyTorch. Installing torch may be required for your platform.
# For large servers, consider installing a suitable torch wheel per https://pytorch.org/
six==1.16.0
python-dotenv==1.0.0
